"half done"

+++ backend/studyapp/src/main/java/com/thesis/studyapp/model/Task.java
@@ -28,7 +28,7 @@ public class Task {
     private int solutionNumber;
 
     private Set<TaskAnswer> answers = new HashSet<>();
-    private int usage;
+    private Long usage;
 
     //todo @JsonIgnoreProperties("createdTasks")?
 //    @JsonIgnore
